Create Redusers. Refactoring Store.
