+++ src/components/Dialogs/Messages/messages.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import s from "./../Dialogs.module.css";
 import Message from "./Message/Message";
-import {addMessageActionCreator, updateNewMessageTextActionCreator} from "../../../redux/State";
+import {addMessageActionCreator, updateNewMessageTextActionCreator} from "../../../redux/dialogs-reducer";
 
 const Messages = (props) => {
