+++ src/redux/store.js
@@ -1,7 +1,6 @@
-const ADD_POST = 'ADD-POST';
-const UPDATE_NEW_POST_TEXT = 'UPDATE-NEW-POST-TEXT';
-const ADD_MESSAGE = 'ADD-MESSAGE';
-const UPDATE_NEW_MESSAGE_TEXT = 'UPDATE-NEW-MESSAGE-TEXT';
+import profileReducer from "./profile-reducer";
+import dialogsReducer from "./dialogs-reducer";
+import usersReducer from "./usersReduser";
 
 let store = {
     _state: {
@@ -30,11 +29,14 @@ let store = {
                 {id: 3, message: 'It is so fun'}
             ]
         },
-        users: [
-            {id: 1, name: 'Serhio'},
-            {id: 2, name: 'Olyalya'},
-            {id: 3, name: 'Dashunya'}
-        ]
+        usersPage: {
+            newUserName: "Yo_Man",
+            users: [
+                {id: 1, name: 'Serhio'},
+                {id: 2, name: 'Olyalya'},
+                {id: 3, name: 'Dashunya'}
+            ]
+        },
     },
     _callSubscriber() {
         console.log('State changed');
@@ -46,45 +48,19 @@ let store = {
         this._callSubscriber = observer
     },
     dispatch(action) {
-        if (action.type === ADD_POST) {
-            let newPost = {
-                id: 5,
-                message: this._state.profilePage.newPostText,
-                likesCount: 0
-            };
-            this._state.profilePage.posts.push(newPost);
-            this._state.profilePage.newPostText = '';
-            this._callSubscriber(this._state);
-        }
-        else if (action.type === UPDATE_NEW_POST_TEXT) {
-            this._state.profilePage.newPostText = action.newText;
-            this._callSubscriber(this._state);
-        }
-        else if (action.type === ADD_MESSAGE) {
-            let newMessage = {
-                id: 4,
-                message: this._state.dialogsPage.newMessageText
-            };
-            this._state.dialogsPage.messages.push(newMessage);
-            this._state.dialogsPage.newMessageText = '';
-            this._callSubscriber(this._state);
-        }
-        else if (action.type === UPDATE_NEW_MESSAGE_TEXT) {
-            this._state.dialogsPage.newMessageText = action.newText;
-            this._callSubscriber(this._state);
-        }
+        this._state.profilePage = profileReducer(this._state.profilePage, action)
+        this._state.DialogsPage = dialogsReducer(this._state.dialogsPage, action)
+        this._state.usersPage = usersReducer(this._state.usersPage, action)
+
+        this._callSubscriber(this._state);
     }
 };
 
-export const addPostActionCreator = () => ({ type: ADD_POST });
-export const updateNewPostTextActionCreator = (text) =>
-    ({ type: UPDATE_NEW_POST_TEXT, newText: text });
-
-export const addMessageActionCreator = () => ({type: ADD_MESSAGE})
-export const updateNewMessageTextActionCreator = (text) =>
-    ({type: UPDATE_NEW_MESSAGE_TEXT, newText: text
-})
 export default store;
+
+
+
+
 
 
 /*addPost() {
